refactor: attachment 엔티티를 글로벌에서 board 도메인으로 이동
