+++ src/main/java/com/ddobang/backend/domain/board/dto/request/AttachmentRequest.java
@@ -1,0 +1,4 @@
+package com.ddobang.backend.domain.board.dto.request;
+
+public class AttachmentRequest {
+}
