+++ src/main/java/com/ddobang/backend/domain/board/repository/AttachmentRepository.java
@@ -3,7 +3,7 @@ package com.ddobang.backend.domain.board.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.ddobang.backend.global.entity.Attachment;
+import com.ddobang.backend.domain.board.entity.Attachment;
 
 @Repository
 public interface AttachmentRepository extends JpaRepository<Attachment, Long> {
