+++ src/test/java/com/ddobang/backend/domain/upload/controller/UploadControllerTest.java
@@ -35,7 +35,7 @@ import com.ddobang.backend.domain.member.entity.Member;
 import com.ddobang.backend.domain.member.repository.MemberRepository;
 import com.ddobang.backend.domain.upload.service.UploadService;
 import com.ddobang.backend.domain.upload.types.FileUploadTarget;
-import com.ddobang.backend.global.entity.Attachment;
+import com.ddobang.backend.domain.board.entity.Attachment;
 import com.ddobang.backend.global.util.Ut;
 
 import jakarta.persistence.EntityManager;
